+++ scripts/director/index.js
@@ -1,2 +1,4 @@
 export { EncounterDirectorApp } from "./encounter-director-app.js";
-export { initializeEncounterDirectorUi, openEncounterDirector, findPreferredEncounterInstanceId } from "./encounter-director-ui.js";
+export { EncounterInstanceManagerApp } from "./encounter-instance-manager-app.js";
+export { initializeEncounterDirectorUi, openEncounterDirector, findPreferredEncounterInstanceId, findEncounterDirectorCandidates } from "./encounter-director-ui.js";
+export { openEncounterInstanceManager } from "./encounter-instance-manager-ui.js";